requirements.txt
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
open-clip-torch>=2.23.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
jupyter>=1.0.0
ipywidgets>=8.0.0
pyyaml>=6.0
tqdm>=4.66.0
scikit-learn>=1.3.0
groq>=0.4.1
python-dotenv>=1.0.0
